[#583] Cleanup + extracted some constants
